describe the algorithm in the README
