+++ isovar/protein_sequence.py
@@ -14,9 +14,10 @@
 
 from __future__ import print_function, division, absolute_import
 
-from collections import namedtuple, defaultdict, OrderedDict
+from collections import namedtuple, OrderedDict
 
 import pandas as pd
+from skbio import DNA
 
 from .logging import create_logger
 from .reference_context import (
@@ -35,11 +36,16 @@ logger = create_logger(__name__)
 # This object combines the information of a ReferenceContext
 # (which establishes the reading frame for particular reference sequence) with
 # a VariantSequence (which contains a cDNA sequence and its supporting reads).
+
 ProteinSequence = namedtuple(
     "ProteinSequence",
     [
+        # translated sequence of a variant sequence in the ORF established
+        # by a reference context
         "amino_acids",
-        "variant_aa_start_offset",
+        # half-open interval coordinates for variant amino acids
+        # in the translated sequence
+        "variant_aa_interval_start",
         "variant_aa_end_offset",
         "reference_context"
         "variant_sequence",
@@ -52,226 +58,255 @@ PROTEIN_FRAGMENT_LEGNTH = 25
 MAX_SEQUENCES_PER_VARIANT = 5
 
 
-def rna_sequence_key(fragment_info):
+def translate(
+        reference_context,
+        cdna_sequence_before_variant,
+        cdna_sequence_of_variant,
+        cdna_sequence_after_variant,
+        transcript_id,
+        transcript_reading_frame_at_start_of_context_sequence,
+        transcript_full_cdna,
+        transcript_full_protein,
+        variant_base1_offset_in_transcript):
     """
-    Get the cDNA (prefix, variant, suffix) fields from a
-    TranslationFromReferenceORF object.
+    Try to  translate the detected cDNA sequence context around a variant using
+    the reading frame of reference transcripts (which must share sequences
+    around the variant).
 
     Parameters
     ----------
-    fragment_info : TranslationFromReferenceORF
+    reference_context : ReferenceContext
+        Has the following fields:
+         - strand
+         - transcript_ids
+         - transcript_names
+         - gene
+         - reference_cdna_sequence_around_variant
+         - reading_frame_at_start_of_context
+         - number_cdna_bases_before_variant
+         - number_cdna_bases_after_variant
+         - reference_protein_sequence_around_variant
 
-    Returns (str, str, str)
+    To determine the reading frame we assume that the reference transcript
+    sequence up to the variant largely matches the sequence we detected
+    from RNA. If this is violated then the returned
+    TranslationFromReferenceContext will be incomplete.
     """
-    return (
-        fragment_info.cdna_prefix,
-        fragment_info.cdna_variant,
-        fragment_info.cdna_suffix
-    )
+    if transcript_reading_frame_at_start_of_context_sequence == 1:
+        # if we're 1 nucleotide into the codon then we need to shift
+        # over two more to restore the ORF
+        orf_offset = 2
+    elif transcript_reading_frame_at_start_of_context_sequence == 2:
+        orf_offset = 1
+    else:
+        orf_offset = 0
 
-def group_protein_fragments(
-        protein_fragment_and_read_names_list,
-        protein_fragment_length=PROTEIN_FRAGMENT_LEGNTH):
+    logger.info("ORF offset into sequence %s_%s_%s from transcript %s: %d" % (
+        cdna_sequence_before_variant,
+        cdna_sequence_of_variant,
+        cdna_sequence_after_variant,
+        transcript_id,
+        orf_offset))
+
+    # translate the variant cDNA sequence we detected from spanning reads
+    # using the ORF offset from the current reference transcript
+    combined_variant_cdna_sequence = DNA(
+        cdna_sequence_before_variant + cdna_sequence_of_variant + cdna_sequence_after_variant)
+
+    variant_protein_fragment_sequence = str(
+        combined_variant_cdna_sequence[orf_offset:].translate())
+
+    logger.info("Combined variant cDNA sequence %s translates to protein %s" % (
+        combined_variant_cdna_sequence,
+        variant_protein_fragment_sequence))
+
+    # translate the reference sequence using the given ORF offset,
+    # we can probably sanity check this by making sure it matches a
+    # substring of the transcript.protein_sequence field
+    combined_transcript_sequence = DNA(
+        transcript_full_cdna[
+            query_sequence_start_idx:
+            query_sequence_start_idx + len(combined_variant_cdna_sequence)])
+
+    transcript_protein_fragment_sequence = str(
+        combined_transcript_sequence[orf_offset].translate())
+
+    fragment_aa_start_offset_in_protein = (
+        query_sequence_start_idx - start_codon_idx) // 3
+    fragment_aa_end_offset_in_protein = (
+        query_sequence_end_idx - start_codon_idx) // 3 + 1
+
+    n_prefix_codons = n_prefix_nucleotides // 3
+
+    if variant_is_deletion and variant_is_frameshift:
+        # after a deletion that shifts the reading frame,
+        # there is a partial codon left, which may be different from
+        # the original codon at that location. Additionally,
+        # the reading frame for all subsequent codons will be different
+        # from the reference
+        variant_codons_start_offset_in_fragment = None
+
+    # TODO: change variant_codons_start_offset_in_fragment in the
+    # TranslationFromReferenceORF to
+    # - variant_amino_acids_start_offset_in_fragment
+    # - variant_amino_acids_end_offset_in_fragment
+    # which are computed from _codons_ offsets by checking which amino
+    # acids differ from the reference sequence(s).
+    # If a mutated sequence ends up having no differences then
+    # we should skip it as a synonymous mutation.
+
+    # TODO #2:
+    # Instead of doing this once per sequence/transcript pair, we
+    # should precompute the reference transcript sequences and the
+    # ORFs they imply and pass in a list of ReferenceTranscriptContext
+    # objects with the following field:
+    #   sequence_before_variant_locus : cDNA
+    #   reading_frame_at_start_of_sequence : int
+    #   transcript_ids : str set
+    #   transcript_names : str set
+    #   gene : str
+
+    variant_aa_start_offset_in_fragment = 0
+    last_variant_codon = 0
+
     """
-    If we end up with multiple equivalent protein fragments from distinct
-    cDNA sequences then we want to group them since ultimately it's
-    not relevant which codon gives us a particular amino acid.
+    # the number of non-mutated codons in the prefix (before the variant)
+    # has to trim the ORF offset and then count up by multiples of 3
 
-    Parameters
-    ----------
-    fragments_and_read_names : list
-        List of tuples containing (1) a TranslationFromReferenceORF object and
-        (2) a set of read names from spanning RNA reads for that
-        unique sequence.
 
-    protein_fragment_length : int
-        Length of the translated protein sequences
+    aa_prefix = variant_protein_fragment_sequence[:n_prefix_codons]
+    aa_variant = variant_protein_fragment_sequence[
+        n_prefix_codons:n_prefix_codons + n_variant_codons]
+    aa_suffix = variant_protein_fragment_sequence[
+        n_prefix_codons + n_variant_codons:]
 
-    Returns list of ProteinFragment objects.
+    assert aa_prefix + aa_variant + aa_suffix == variant_protein_fragment_sequence
+                    aa_prefix=aa_prefix,
+            aa_variant=aa_variant,
+            aa_suffix=aa_suffix)
     """
-    # map every distinct protein sequence to a tuple with the following fields:
-    # - transcript ID
-    # - transcript name
-    # - cDNA sequence
-    # - reading frame
 
-    protein_sequence_dict = defaultdict(list)
-    cdna_to_read_names = {}
-    for (translation_info, read_names) in protein_fragment_and_read_names_list:
-        cnda_tuple = rna_sequence_key(translation_info)
-        if cnda_tuple in cdna_to_read_names:
-            assert cdna_to_read_names[cnda_tuple] == read_names
-        else:
-            cdna_to_read_names[cnda_tuple] = read_names
 
-        protein_sequence = translation_info.variant_protein_sequence
-        # if we can chop up the translated protein fragment into shorter pieces
-        # then go for it!
 
-        variant_protein_sequence = translation_info.variant_protein_sequence
-        ref_protein_sequence = translation_info.reference_protein_sequence
-        n_amino_acids = len(variant_protein_sequence)
-
-        for start, end in enumerate(range(protein_fragment_length, n_amino_acids)):
-
-            # all protein fragments must overlap the variant
-            if start > translation_info.base0_variant_amino_acid_start_offset:
-                break
-
-            if end < translation_info.base0_variant_amino_acid_end_offset:
-                continue
-
-            variant_protein_subsequence = protein_sequence[start:end]
-            ref_protein_subsequence = ref_protein_sequence[start:end]
-            info_tuple = (
-                variant_protein_subsequence,
-                ref_protein_subsequence,
-                translation_info.transcript_id,
-                translation_info.transcript_name,
-                translation_info.base0_variant_amino_acid_start_offset,
-                translation_info.base0_variant_amino_acid_end_offset,
-            )
-            protein_sequence_dict[protein_sequence].append(info_tuple)
-
-    results = []
-    for (protein_sequence, info_objs) in protein_sequence_dict.items():
-        transcript_names = list(set([x.transcript_name for x in info_objs]))
-        transcript_ids = list(set([x.transcript_id for x in info_objs]))
-        cdna_sequence_keys = list(set([rna_sequence_key(x) for x in info_objs]))
-
-        reference_protein_sequences = []
-
-        total_read_count = sum(
-            len(cdna_to_read_names[cnda_tuple])
-            for cnda_tuple in cdna_sequence_keys)
-        transcript_ids = list(set([x.transcript_id for x in info_objs]))
-        transcript_names = list(set([x.transcript_name for x in info_objs]))
-        reference_protein_sequences = list(set([
-            x.reference_protein_sequence
-            for x in info_objs]))
-        cdna_sequences_to_read_names = {
-            rna_sequence_key(x): cdna_to_read_names[rna_sequence_key(x)]
-            for x in info_objs
-        }
-        list(set(
-            [rna_sequence_key(x) for x in info_objs]))
-        results.append(
-            ProteinSequence(
-                number_supporting_reads=total_read_count,
-                variant_protein_sequence=protein_sequence,
-                reference_transcript_ids=transcript_ids,
-                reference_transcript_names=transcript_names,
-                reference_protein_sequences=reference_protein_sequences,
-                cdna_sequences_to_read_names=cdna_sequences_to_read_names))
-    return results
-
-def translate_variant(
-        variant,
-        sequence_context_to_read_names_dict,
-        variant_nucleotides,
-        reference_transcripts,
-        protein_fragment_length=PROTEIN_FRAGMENT_LEGNTH,
-        min_reads_supporting_rna_sequence=MIN_READS_SUPPORTING_RNA_SEQUENCE,
-        min_transcript_prefix_length=MIN_TRANSCRIPT_PREFIX_LENGTH,
+def translate_variant_on_transcript(
+        dna_sequence_prefix,
+        dna_sequence_variant,
+        dna_sequence_suffix,
+        variant_is_insertion,
+        variant_is_deletion,
+        variant_is_frameshift,
+        base1_variant_start_location,
+        base1_variant_end_location,
+        transcript,
         max_transcript_mismatches=MAX_TRANSCRIPT_MISMATCHES,
-        max_sequences=MAX_SEQUENCES_PER_VARIANT):
+        min_transcript_prefix_length=MIN_TRANSCRIPT_PREFIX_LENGTH):
+    if transcript.strand == "+":
+        cdna_prefix = dna_sequence_prefix
+        cdna_suffix = dna_sequence_suffix
+        cdna_variant = dna_sequence_variant
+    else:
+        # if the transcript is on the reverse strand then we have to
+        # take the sequence PREFIX|VARIANT|SUFFIX
+        # and take the complement of XIFFUS|TNAIRAV|XIFERP
+        cdna_prefix = str(DNA(dna_sequence_suffix).reverse_complement())
+        cdna_suffix = str(DNA(dna_sequence_prefix).reverse_complement())
+        cdna_variant = str(
+            DNA(dna_sequence_variant).reverse_complement())
+
+    n_mismatch_before_variant = sum(
+        xi != yi
+        for (xi, yi) in zip(
+            transcript_sequence_before_variant, cdna_prefix))
+
+    if n_mismatch_before_variant > max_transcript_mismatches:
+        logger.info(
+            "Skipping transcript %s, too many mismatching bases (%d)",
+            transcript,
+            n_mismatch_before_variant)
+        return None
+
+    transcript_reading_frame = (query_sequence_start_idx - start_codon_idx) % 3
+
+    return TranslationFromReferenceORF(
+        cdna_prefix=cdna_prefix,
+        cdna_variant=cdna_variant,
+        cdna_suffix=cdna_suffix,
+        transcript_id=transcript.id,
+        transcript_name=transcript.name,
+        number_transcript_sequence_mismatches=n_mismatch,
+        reading_frame_at_start_of_cdna_sequence=reading_frame,
+        transcript_sequence_before_variant=transcript_sequence_before_variant,
+        variant_protein_sequence=variant_protein_fragment_sequence,
+        reference_protein_sequence=transcript_protein_fragment_sequence,
+        fragment_aa_start_offset_in_protein=fragment_aa_start_offset_in_protein,
+        fragment_aa_end_offset_in_protein=fragment_aa_end_offset_in_protein,
+        variant_aa_start_offset_in_fragment=variant_aa_start_offset_in_fragment,
+        variant_aa_end_offset_in_fragment=variant_aa_end_offset_in_fragment)
+
+def translate_variant_sequence(
+        variant_sequence,
+        reference_context,
+        max_transcript_mismatches):
     """
-    Generate possible protein sequences around a variant from surrounding
-    context sequences and a set of reference transcripts from the same locus
-    which can be used to establish an ORF.
+    Attempt to translate a single VariantSequence using the reading frame
+    from a single ReferenceContext.
 
     Parameters
     ----------
-    variant : varcode.Variant
+    variant_sequence : VariantSequence
 
-    sequence_context_to_read_names_dict : dict
-        Dictionary mapping pair of (prefix, suffix) sequences to a set of
-        read names which support this sequence.
+    reference_context : ReferenceContext
 
-    variant_nucleotides : str
-        Nucleotides found between the (prefix, suffix) pair of each context.
+    max_transcript_mismatches : int
+        Don't use the reading frame from a context where the cDNA variant
+        sequences disagrees at more than this number of positions before the
+        variant nucleotides.
 
-    reference_transcripts : list of pyensembl.Transcript
-        Transcripts which are used to establish the reading frame for
-        the discovered cDNA sequences.
+    Returns either a ProteinSequence object or None.
+    """
+    pass
 
-    protein_fragment_length : int
 
-    min_reads_supporting_rna_sequence : int
+def translate_multiple_protein_sequences_from_variant_sequences(
+        variant_sequences,
+        reference_contexts,
+        max_transcript_mismatches,
+        max_protein_sequences_per_variant):
+    """
+    Given all detected VariantSequence objects for a particular variant
+    and all the ReferenceContext objects for that locus, translate
+    multiple protein sequences, up to the number specified by the argument
+    max_protein_sequences_per_variant.
 
-    min_transcript_prefix_length : int
+    Parameters
+    ----------
+    variant_sequences : list of VariantSequence objects
+        Variant sequences overlapping a single original variant
+
+    reference_contexts : list of ReferenceContext objects
+        Reference sequence contexts from the same variant as the variant_sequences
 
     max_transcript_mismatches : int
 
-    max_sequences : int
-
-    chromosome_name : str, optional
-        If given, then use this name instead of the chromosome name on
-        the variant object (necessary when the BAM's reference is hg19)
-
-    Returns list of tuples containing:
-        1) ProteinFragment object
-        2) set of read names supporting the protein fragment
+    max_protein_sequences_per_variant
     """
-    protein_fragments_and_read_names = []
-
-    for i, ((prefix, suffix), read_names) in enumerate(sorted(
-            sequence_context_to_read_names_dict.items(),
-            key=lambda x: -len(x[1]))):
-        logger.info("%d %s|%s %s" % (
-            i,
-            prefix,
-            suffix,
-            read_names))
-
-        if i >= max_sequences:
-            logger.info(
-                "Skipping sequence %s for variant %s, already reached max_sequences (%d)",
-                prefix + "|" + suffix,
-                variant,
-                max_sequences)
-            break
-
-        num_reads_supporting_current_sequence = len(read_names)
-
-        if num_reads_supporting_current_sequence < min_reads_supporting_rna_sequence:
-            logger.info(
-                "Skipping sequence %s for variant %s, too few supporting reads (%d)",
-                prefix + "|" + suffix,
-                variant,
-                num_reads_supporting_current_sequence)
-            # we can break here instead of `continue` since the loop iterations
-            # are sorted in decreasing order by the number of reads
-            break
-
-        for protein_fragment in translate_compatible_reading_frames(
-                dna_sequence_prefix=prefix,
-                dna_sequence_variant=variant_nucleotides,
-                dna_sequence_suffix=suffix,
-                base1_variant_start_location=variant.start,
-                base1_variant_end_location=variant.end,
-                variant_is_insertion=variant.is_insertion,
-                variant_is_deletion=variant.is_deletion,
-                variant_is_frameshift=not variant.preserves_reading_frame,
-                transcripts=reference_transcripts,
-                max_transcript_mismatches=max_transcript_mismatches,
-                min_transcript_prefix_length=max_transcript_mismatches):
-            protein_fragments_and_read_names.append(
-                (protein_fragment, read_names))
-
-    logger.info("Gathered protein fragments for %s: %s" % (
-        variant,
-        protein_fragments_and_read_names))
-
-    grouped_results = group_protein_fragments(
-        protein_fragments_and_read_names,
-        protein_fragment_length=protein_fragment_length)
-
-    logger.info("Grouped protein fragments for %s: %s" % (
-        variant,
-        grouped_results))
-    return grouped_results
-
+    protein_sequences = []
+    for reference_context in reference_contexts:
+        for variant_sequence in variant_sequences:
+            # assuming that the reference contexts and variant sequences
+            # were sorted in order of priority, we will also get back
+            # translations in order of quality, so it's OK to only keep
+            # the top k
+            n_protein_sequences = len(protein_sequences)
+            if n_protein_sequences >= max_protein_sequences_per_variant:
+                return protein_sequences
+            protein_sequence = translate_variant_sequence(
+                variant_sequence=variant_sequence,
+                reference_context=reference_context,
+                max_transcript_mismatches=max_transcript_mismatches)
+            if protein_sequence is not None:
+                protein_sequences.append(protein_sequence)
+    return protein_sequences
 
 def translate_variants(
         variants,
@@ -333,6 +368,9 @@ def translate_variants(
             samfile=samfile,
             sequence_length=rna_sequence_length,
             min_reads=min_reads_supporting_rna_sequence):
+        # include every variant in the dictionary, even if it doesn't yield any
+        # translated proteins, just in case we later want to count how many
+        # translations succeeded vs. failed
         variant_to_protein_sequences_dict[variant] = []
 
         if len(variant_sequences) == 0:
@@ -360,24 +398,12 @@ def translate_variants(
             variant,
             context_size=context_size,
             transcript_id_whitelist=transcript_id_whitelist)
-
-        for reference_context in reference_contexts:
-            for variant_sequence in variant_sequences:
-                translations = translate_variant_sequence(
-                    variant_sequence=variant_sequence,
-                    reference_context=reference_context,
-                    max_transcript_mismatches=max_transcript_mismatches,
-                    protein_fragment_length=protein_fragment_length)
-                for translation in translations:
-                    # assuming that the reference contexts and variant sequences
-                    # were sorted in order of priority, we will also get back
-                    # translations in order of quality, so it's OK to only keep
-                    # the top k
-                    n_protein_sequences = len(variant_to_protein_sequences_dict[variant])
-                    if n_protein_sequences >= max_protein_sequences_per_variant:
-                        break
-                    else:
-                        variant_to_protein_sequences_dict[variant].append(translation)
+        variant_to_protein_sequences_dict[variant] = \
+            translate_multiple_protein_sequences_from_variant_sequences(
+                variant_sequences=variant_sequences,
+                reference_contexts=reference_contexts,
+                max_transcript_mismatches=max_transcript_mismatches,
+                max_protein_sequences_per_variant=max_protein_sequences_per_variant)
     return variant_to_protein_sequences_dict
 
 def translate_variants_dataframe(
